+++ tests-run-default.sh
@@ -20,12 +20,13 @@ echo "CARGO TEST (release)"
 cargo test --release
 
 echo
-echo "CARGO TEST (debug, feature lint_unused_unsafe)"
-cargo test --features lint_unused_unsafe
-# feature "lint_unused_unsafe" can't be tested/used in release, only in debug (above)
+echo "CARGO CHECK --TESTS (debug, feature lint_unused_unsafe)"
+cargo check --tests --features lint_unused_unsafe
+# - feature "lint_unused_unsafe" can't be tested/built, only checked; and not in release, only in
+#   debug
 
 echo
-echo "CARGO TEST (negative_tests/verify_error_messages, debug, feature unused_lint)"
+echo "CARGO CHECK --TESTS (negative_tests/verify_error_messages, debug, feature unused_lint)"
 cd negative_tests/verify_error_messages
-cargo test --features lint_unused_unsafe
+cargo check --tests --features lint_unused_unsafe
 cd - >/dev/null